Invite + Permission work

+++ src/main/java/de/marcschuler/webrtcserver/controller/v0/ChannelController.java
@@ -1,9 +1,11 @@
 package de.marcschuler.webrtcserver.controller.v0;
 
+import de.marcschuler.webrtcserver.data.permission.PermissionType;
 import de.marcschuler.webrtcserver.dto.ChannelCreateDTO;
 import de.marcschuler.webrtcserver.dto.data.ChannelDTO;
 import de.marcschuler.webrtcserver.mapper.ServerMapper;
 import de.marcschuler.webrtcserver.service.ChannelService;
+import de.marcschuler.webrtcserver.service.PermissionService;
 import de.marcschuler.webrtcserver.service.SectionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
@@ -20,10 +22,12 @@ public class ChannelController implements OrderableController {
     private final SectionService sectionService;
     private final ChannelService channelService;
 
+    private final PermissionService permissionService;
+
     private final ServerMapper serverMapper;
 
     @GetMapping("{id}")
-    public ChannelDTO channel(@PathVariable UUID id){
+    public ChannelDTO channel(@PathVariable UUID id) {
         return channelService.get(id).map(serverMapper::mapToDTO)
                 .orElseThrow();
     }
@@ -31,6 +35,8 @@ public class ChannelController implements OrderableController {
     @PostMapping
     public ChannelDTO create(@RequestBody ChannelCreateDTO channelCreateDTO) {
         var section = sectionService.get(channelCreateDTO.getSectionId()).orElseThrow();
+        permissionService.checkControllerAccess(section, null, PermissionType.SECTION_CHANNEL_CREATE);
+
         var channel = channelService.create(channelCreateDTO.getName(), section);
         return serverMapper.mapToDTO(channel);
     }
@@ -38,6 +44,9 @@ public class ChannelController implements OrderableController {
     @PutMapping("{id}")
     public ChannelDTO edit(@PathVariable UUID id, @RequestBody ChannelDTO channelDTO) {
         var channel = channelService.get(id).orElseThrow();
+
+        permissionService.checkControllerAccess(channel, PermissionType.CHANNEL_EDIT);
+
         channelService.edit(channel, channelDTO);
         return serverMapper.mapToDTO(channel);
     }
@@ -45,6 +54,9 @@ public class ChannelController implements OrderableController {
     @Override
     public void order(UUID id, int newOrder) {
         var channel = channelService.get(id).orElseThrow();
+
+        permissionService.checkControllerAccess(channel, PermissionType.SECTION_CHANNEL_ORDER);
+
         channelService.order(channel, newOrder);
     }
 
@@ -52,12 +64,22 @@ public class ChannelController implements OrderableController {
     public void move(@PathVariable UUID id, @PathVariable UUID newParentId, @PathVariable int newOrder) {
         var channel = channelService.get(id).orElseThrow();
         var newSection = sectionService.get(newParentId).orElseThrow();
+
+        // Check both old and new section for move permissions
+        permissionService.checkControllerAccess(channel.getSection(),channel, PermissionType.SECTION_CHANNEL_MOVE);
+        permissionService.checkControllerAccess(newSection, channel,PermissionType.SECTION_CHANNEL_MOVE);
+
+        permissionService.checkControllerAccess(channel, PermissionType.CHANNEL_EDIT);
+
         channelService.move(channel, newSection, newOrder);
     }
 
     @DeleteMapping("{id}")
     public void delete(@PathVariable UUID id) {
         var channel = channelService.get(id).orElseThrow();
+
+        permissionService.checkControllerAccess(channel, PermissionType.CHANNEL_DELETE);
+
         channelService.delete(channel);
     }
 }
