+++ src/main/java/de/marcschuler/webrtcserver/controller/v0/ServerController.java
@@ -1,9 +1,11 @@
 package de.marcschuler.webrtcserver.controller.v0;
 
+import de.marcschuler.webrtcserver.data.permission.PermissionType;
 import de.marcschuler.webrtcserver.dto.data.ServerDTO;
 import de.marcschuler.webrtcserver.dto.data.message.MessageContentDTO;
 import de.marcschuler.webrtcserver.mapper.MessageContentMapper;
 import de.marcschuler.webrtcserver.mapper.ServerMapper;
+import de.marcschuler.webrtcserver.service.PermissionService;
 import de.marcschuler.webrtcserver.service.ServerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
@@ -18,12 +20,15 @@ import java.util.UUID;
 public class ServerController {
 
     private final ServerService serverService;
+    private final PermissionService permissionService;
 
     private final ServerMapper serverMapper;
     private final MessageContentMapper messageContentMapper;
 
     @PutMapping("{serverId}")
     public ServerDTO edit(@PathVariable UUID serverId, @RequestBody ServerDTO serverDto) {
+        permissionService.checkControllerAccess(null, null, PermissionType.SERVER_EDIT);
+
         var server = serverService.get(serverId).orElseThrow();
         server = serverService.update(server, serverDto);
         return serverMapper.mapToDTO(server);
@@ -43,6 +48,8 @@ public class ServerController {
 
         @PostMapping
         public MessageContentDTO create(@RequestBody MessageContentDTO messageDto, @PathVariable UUID id) {
+            permissionService.checkControllerAccess(null, null, PermissionType.SERVER_EDIT_DESCRIPTION);
+
             var server = serverService.get(id).orElseThrow();
             var content = serverService.createDescription(server, messageDto);
             return messageContentMapper.mapToDTO(content);
@@ -50,6 +57,8 @@ public class ServerController {
 
         @PutMapping("{descriptionId}")
         public MessageContentDTO edit(@RequestBody MessageContentDTO messageDto, @PathVariable UUID id, @PathVariable UUID descriptionId) {
+            permissionService.checkControllerAccess(null, null, PermissionType.SERVER_EDIT_DESCRIPTION);
+
             var server = serverService.get(id).orElseThrow();
             var content = serverService.descriptionFromId(server, descriptionId).orElseThrow();
             content = serverService.updateDescription(server, content, messageDto);
@@ -58,6 +67,8 @@ public class ServerController {
 
         @PutMapping("{descriptionId}/order/{newOrder}")
         public void order(@PathVariable UUID id, @PathVariable UUID descriptionId, @PathVariable int newOrder) {
+            permissionService.checkControllerAccess(null, null, PermissionType.SERVER_EDIT_DESCRIPTION);
+
             var server = serverService.get(id).orElseThrow();
             var content = serverService.descriptionFromId(server, descriptionId).orElseThrow();
             serverService.orderDescription(server, content, newOrder);
@@ -65,6 +76,8 @@ public class ServerController {
 
         @DeleteMapping("{descriptionId}")
         public void delete(@PathVariable UUID id, @PathVariable UUID descriptionId) {
+            permissionService.checkControllerAccess(null, null, PermissionType.SERVER_EDIT_DESCRIPTION);
+
             var server = serverService.get(id).orElseThrow();
             serverService.deleteDescription(server, descriptionId);
         }
